Add 3rd-party-allowed mod downloading ability

+++ src/json/manifest.rs
@@ -6,7 +6,7 @@ pub struct JsonManifest {
     pub minecraft: JsonMinecraft,
     pub manifest_type: String,
     pub overrides: String,
-    pub manifest_version: i64,
+    pub manifest_version: u32,
     pub version: String,
     pub author: String,
     pub name: String,
@@ -31,8 +31,8 @@ pub struct JsonModLoader {
 #[serde(rename_all = "camelCase")]
 pub struct JsonFile {
     #[serde(rename = "projectID")]
-    pub project_id: i64,
+    pub project_id: u32,
     #[serde(rename = "fileID")]
-    pub file_id: i64,
+    pub file_id: u32,
     pub required: bool,
 }
